Refactored subscribe and server
