+++ litmus-portal/cluster-agents/subscriber/pkg/graphql/operations.go
@@ -1,0 +1,44 @@
+package graphql
+
+import (
+	"bytes"
+	"encoding/json"
+	"io/ioutil"
+	"net/http"
+	"strconv"
+	"strings"
+)
+
+func SendRequest(server string, payload []byte) (string, error) {
+	req, err := http.NewRequest("POST", server, bytes.NewBuffer(payload))
+	if err != nil {
+		return "", err
+	}
+	req.Header.Set("Content-Type", "application/json")
+	client := &http.Client{}
+	resp, err := client.Do(req)
+	if err != nil {
+		return "", err
+	}
+
+	body, err := ioutil.ReadAll(resp.Body)
+	resp.Body.Close()
+	if err != nil {
+		return "", err
+	}
+
+	return string(body), nil
+}
+
+// process event data into proper format acceptable by graphql
+func MarshalGQLData(gqlData interface{}) (string, error) {
+	data, err := json.Marshal(gqlData)
+	if err != nil {
+		return "", err
+	}
+
+	// process the marshalled data to make it graphql compatible
+	processed := strconv.Quote(string(data))
+	processed = strings.Replace(processed, `\"`, `\\\"`, -1)
+	return processed, nil
+}
