+++ litmus-portal/cluster-agents/subscriber/pkg/types/connect.go
@@ -48,8 +48,8 @@ type KubeGVRRequest struct {
 }
 
 type Action struct {
-	K8SManifest  string      `json:"k8s_manifest"`
-	ExternalData string 	 `json:"external_data"`
-	RequestType  string      `json:"request_type"`
-	Namespace    string      `json:"namespace"`
+	K8SManifest  string `json:"k8s_manifest"`
+	ExternalData string `json:"external_data"`
+	RequestType  string `json:"request_type"`
+	Namespace    string `json:"namespace"`
 }
