+++ litmus-portal/cluster-agents/subscriber/pkg/workflow/util.go
@@ -2,21 +2,20 @@ package workflow
 
 import (
 	"errors"
-	"github.com/litmuschaos/litmus/litmus-portal/cluster-agents/subscriber/pkg/gql"
-	"github.com/litmuschaos/litmus/litmus-portal/cluster-agents/subscriber/pkg/k8s"
-	"github.com/sirupsen/logrus"
-	"regexp"
-	"strconv"
-	"strings"
-	"time"
-
 	"github.com/argoproj/argo/pkg/apis/workflow/v1alpha1"
 	v1alpha13 "github.com/argoproj/argo/pkg/apis/workflow/v1alpha1"
 	wfclientset "github.com/argoproj/argo/pkg/client/clientset/versioned"
 	v1alpha12 "github.com/litmuschaos/chaos-operator/pkg/client/clientset/versioned/typed/litmuschaos/v1alpha1"
+	"github.com/litmuschaos/litmus/litmus-portal/cluster-agents/subscriber/pkg/graphql"
+	"github.com/litmuschaos/litmus/litmus-portal/cluster-agents/subscriber/pkg/k8s"
 	"github.com/litmuschaos/litmus/litmus-portal/cluster-agents/subscriber/pkg/types"
+	"github.com/sirupsen/logrus"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"regexp"
+	"strconv"
+	"strings"
+	"time"
 
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/runtime/serializer/yaml"
@@ -126,6 +125,7 @@ func WorkflowRequest(clusterData map[string]string, requestType string, uid stri
 			return err
 		}
 
+		logrus.Info("workflow delete name: ", wfOb.Name, "namespace: ", wfOb.Namespace)
 	} else if requestType == "workflow_sync" {
 		wfOb, err := GetWorkflowObj(uid)
 		if err != nil {
@@ -136,10 +136,11 @@ func WorkflowRequest(clusterData map[string]string, requestType string, uid stri
 
 		events, err := WorkflowEventHandler(wfOb, "UPDATE", startTime)
 		if err != nil {
+			logrus.Info(err)
 			return err
 		}
 
-		response, err := gql.SendWorkflowUpdates(clusterData, events)
+		response, err := SendWorkflowUpdates(clusterData, events)
 		if err != nil {
 			return err
 		}
@@ -148,14 +149,14 @@ func WorkflowRequest(clusterData map[string]string, requestType string, uid stri
 	return nil
 }
 
-func GetWorkflowObj(uid string) (*v1alpha1.Workflow, error){
+func GetWorkflowObj(uid string) (*v1alpha1.Workflow, error) {
 	conf, err := k8s.GetKubeConfig()
 	if err != nil {
 		return nil, err
 	}
 
 	// create the workflow client
-	wfClient := wfclientset.NewForConfigOrDie(conf).ArgoprojV1alpha1().Workflows("default")
+	wfClient := wfclientset.NewForConfigOrDie(conf).ArgoprojV1alpha1().Workflows(AgentNamespace)
 	listWf, err := wfClient.List(metav1.ListOptions{})
 	if err != nil {
 		return nil, err
@@ -177,6 +178,24 @@ func DeleteWorflow(wfname string) error {
 	}
 
 	// create the workflow client
-	wfClient := wfclientset.NewForConfigOrDie(conf).ArgoprojV1alpha1().Workflows("default")
+	wfClient := wfclientset.NewForConfigOrDie(conf).ArgoprojV1alpha1().Workflows(AgentNamespace)
 	return wfClient.Delete(wfname, &metav1.DeleteOptions{})
+}
+
+// generate graphql mutation payload for workflow event
+func GenerateWorkflowPayload(cid, accessKey, completed string, wfEvent types.WorkflowEvent) ([]byte, error) {
+	clusterID := `{cluster_id: \"` + cid + `\", access_key: \"` + accessKey + `\"}`
+
+	for id, event := range wfEvent.Nodes {
+		event.Message = strings.Replace(event.Message, `"`, ``, -1)
+		wfEvent.Nodes[id] = event
+	}
+
+	processed, err := graphql.MarshalGQLData(wfEvent)
+	if err != nil {
+		return nil, err
+	}
+	mutation := `{ workflow_id: \"` + wfEvent.WorkflowID + `\", workflow_run_id: \"` + wfEvent.UID + `\", completed: ` + completed + `, workflow_name:\"` + wfEvent.Name + `\", cluster_id: ` + clusterID + `, execution_data:\"` + processed[1:len(processed)-1] + `\"}`
+	var payload = []byte(`{"query":"mutation { chaosWorkflowRun(workflowData:` + mutation + ` )}"}`)
+	return payload, nil
 }
