+++ litmus-portal/graphql-server/pkg/cluster/handler/handler.go
@@ -225,12 +225,14 @@ func SendRequestToSubscriber(subscriberRequest clusterOps.SubscriberRequests, r 
 			for CreateChaosWorkflow mutation to be passed to this function.
 		*/
 	}
+
 	newAction := &model.ClusterAction{
 		ProjectID: subscriberRequest.ProjectID,
 		Action: &model.ActionPayload{
-			K8sManifest: subscriberRequest.K8sManifest,
-			Namespace:   subscriberRequest.Namespace,
-			RequestType: subscriberRequest.RequestType,
+			K8sManifest:  subscriberRequest.K8sManifest,
+			Namespace:    subscriberRequest.Namespace,
+			RequestType:  subscriberRequest.RequestType,
+			ExternalData: subscriberRequest.ExternalData,
 		},
 	}
